% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplefile.R
\name{coxph_wrapper}
\alias{coxph_wrapper}
\title{A simple wrapper around coxph}
\usage{
coxph_wrapper(...)
}
\arguments{
\item{...}{Arguments to be passed to copxh}
}
\value{
A copxh.object
}
\description{
A simple wrapper around coxph
}

